refactor of opengl api: no need to keep track of the state anymore
the openglRenderApi class keeps track of which element is bound at all times, and only actually calls opengl if the object is not already bound. With this, we can call the internal bind() functions before every operations that alters state. And with that, we don't need public bind methods anymore (only for the framebuffer, because it modifies the viewport as well)